pyproject.toml
[tool.poetry]
name = "ai-microservice"
version = "0.1.0"
description = "Upscaling and downscaling service"
authors = ["shanudengre82 <shanudengre82@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
# Shared dependencies (used by both)
pillow = ">=10.0.0" 
uvicorn = "^0.27.0"
streamlit = ">=1.31.0"
fastapi = "^0.110.0"
requests = ">=2.31.0"
streamlit-image-comparison = ">=0.0.4"
python-multipart = "^0.0.9"
opencv-python-headless = "^4.12.0.88"


[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"